Clean up preamble by removing unnecessary things
Remove various options that aren't needed as they are default.
Flushbottom is already the default for two sided reports (which is
the class that ip3thesis is based on), while for one sided reports
raggedbottom is set, which seems perfectly sensible.
Same for the float locations - these are already the defaults.